Add files via upload
This data set contains information on approximately 12,000 commercially available wines.
The variables are mostly related to the chemical properties of the wine being sold. The
target variable is the number of sample cases of wine that were purchased by wine
distribution companies after sampling a wine. These cases would be used to provide tasting
samples to restaurants and wine stores around the United States. The more sample cases
purchased, the more likely a wine is to be sold at a high end restaurant.

A large wine manufacturer is studying the data in order to predict the number of wine
cases ordered based upon the wine characteristics. If the wine manufacturer can predict
the number of cases, then that manufacturer will be able to adjust their wine offering to
maximize sales.